[CSL-1081] Split 'Pos.Block.Logic' into smaller modules
